+++ changelog.md
@@ -93,7 +93,9 @@
 
 ## Standard library additions and changes
 
-- Added support for parenthesized expressions in `strformat`
+- `strformat`:
+  added support for parenthesized expressions.
+  added support for const string's instead of just string literals
 
 - Fixed buffer overflow bugs in `net`
 
@@ -101,7 +103,9 @@
 
 - Added `sections` iterator in `parsecfg`.
 
-- Make custom op in macros.quote work for all statements.
+- `macros`:
+  Make custom op in `quote` work for all statements.
+  Undeprecate `callsite`.
 
 - On Windows the SSL library now checks for valid certificates.
   It uses the `cacert.pem` file for this purpose which was extracted